move global tests out of lib
