+++ tokio-executor/src/global.rs
@@ -221,23 +221,4 @@ unsafe fn hide_lt<'a>(p: *mut (dyn Executor + 'a)) -> *mut (dyn Executor + 'stat
     mem::transmute(p)
 }
 
-#[cfg(test)]
-mod tests {
-    use super::{with_default, DefaultExecutor};
 
-    #[test]
-    fn default_executor_is_send_and_sync() {
-        fn assert_send_sync<T: Send + Sync>() {}
-
-        assert_send_sync::<DefaultExecutor>();
-    }
-
-    #[test]
-    #[should_panic]
-    fn nested_default_executor_status() {
-        let _enter = super::super::enter().unwrap();
-        let mut executor = DefaultExecutor::current();
-
-        let _result = with_default(&mut executor, || ());
-    }
-}
